mds: remove inc array from mdsmap
Instead, use the MDSMap epochs to provide the `inc` that
we populate into mds_info_t for use by the daemons.

It is not necessary to respect the current value of inc
in old maps, because the per-rank counters were necessarily
lower numbers than the overall map epoch (incrementing
the counters always corresponded to a map change).

Fixes: #15399
Signed-off-by: John Spray <john.spray@redhat.com>